Changes to deploy on netlify

+++ src/utils/storage/localStorage.ts
@@ -5,13 +5,22 @@ export enum LS_KEYS {
 }
 
 export const setData = (key: LS_KEYS, value: object) => {
+    if (typeof localStorage === 'undefined') {
+        return null;
+    }
     localStorage.setItem(key, JSON.stringify(value));
 }
 
 export const getData = (key: LS_KEYS) => {
+    if (typeof localStorage === 'undefined') {
+        return null;
+    }
     return JSON.parse(localStorage.getItem(key));
 }
 
 export const clearData = () => {
+    if (typeof localStorage === 'undefined') {
+        return null;
+    }
     localStorage.clear();
 }
